Implement Postlogistics Cash on Delivery labels
The webservice awaits the cash on delivery amount.
When a whole sales order is delivered in one picking / one package, we the
amount is the total of the sales order. If it is split in several pickings /
packages, we don't know the value of each package and so we ask the user to
fill the price on the packages.